# openapi/main.yml
openapi: 3.0.3
info:
  title:  Rest API
  description: All Item Rest API End Points
  version: V-2024-04
servers:
  - url: http://localhost:8080
tags:
  - name: Note
  - name: Tag
paths:
  /note:
    post:
      tags:
        - Note
      summary: Create a new note
      description: Add a new note in the app
      operationId: createNote
      requestBody:
        description: note data required to create a new note
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteDTO'
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                format: UUID
    get:
      tags:
        - Note
      summary: Get all notes by status
      description: Get all notes by status
      operationId: getAllNotesByStatus
      parameters:
        - name: extractCode
          in: query
          required: true
          description: the code which specifies the number of fields you wish to retrieve
          schema:
            type: string
        - name: noteStatus
          in: query
          required: true
          description: the status of the notes to be retrieved
          schema:
            $ref: '#/components/schemas/NoteStatusDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteDTO'
  /note/{noteId}:
    put:
      tags:
        - Note
      summary: update a note
      description: update a note
      operationId: updateNoteById
      parameters:
        - name: noteId
          in: path
          description: ID of the note to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: new data of the note to modify
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteDTO'
      responses:
        204:
          description: successful operation
    get:
      tags:
        - Note
      summary: get Note by ID
      description: get Note by ID
      operationId: getNoteById
      parameters:
        - name: noteId
          in: path
          description: identifier of the note to obtain
          required: true
          schema:
            type: string
            format: uuid
        - name: extractCode
          in: query
          required: true
          description: the code which specifies the number of fields you wish to retrieve
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDTO'
    delete:
      tags:
        - Note
      summary: delete note by ID
      description: delete note by ID
      operationId: deleteNoteByID
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: successfully operation
  /tag:
    post:
      tags:
        - Tag
      summary: Create a new tag
      description: Create a new tag
      operationId: createTag
      requestBody:
        description: tag data required to create a new tag
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagDTO'
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                format: UUID
    get:
      tags:
        - Tag
      summary: Get all Tag
      description: Get all Tag
      operationId: getAllTag
      parameters:
        - name: extractCode
          in: query
          required: true
          description: the code which specifies the number of fields you wish to retrieve
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagDTO'
  /tag/{tagId}:
    get:
      tags:
        - Tag
      summary: get Tag by ID
      description: get Tag by ID
      operationId: getTagById
      parameters:
        - name: tagId
          in: path
          description: identifier of the tag to obtain
          required: true
          schema:
            type: string
            format: uuid
        - name: extractCode
          in: query
          required: true
          description: the code which specifies the number of fields you wish to retrieve
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagDTO'
  /note/{noteId}/tag/{tagId}:
    put:
      tags:
        - Note
      summary: add a Tag to a Note
      description: add a Tag to a Note
      operationId: addTagToNoteById
      parameters:
        - name: noteId
          in: path
          required: true
          description: identifier of the Note in which we will add the Tag
          schema:
            type: string
            format: uuid
        - name: tagId
          in: path
          required: true
          description: identifier of the Tag
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: successful operation
components:
  schemas:
    CreateNoteDTO:
      type: object
      properties:
        title:
          type: string
          description: title of the note
        content:
          type: string
          description: content of the note
    CreateTagDTO:
      type: object
      properties:
        name:
          type: string
          description: name of the tag
        description:
          type: string
          description: description of the tag
    UpdateNoteDTO:
      type: object
      properties:
        title:
          type: string
          description: title of the note
        content:
          type: string
          description: content of the note
    NoteDTO:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    NoteStatusDTO:
      type: string
      enum:
        - DRAFT
        - DELETED
        - PUBLISHED
    TagDTO:
      type: object
      properties:
        name:
          type: string
          description: name of the tag
        description:
          type: string
          description: description of the tag